changed dynamic routes to useRouter from getStaticPaths
Because getStaticPaths is, well, static, when a user created a new debt it would not register as a valid url path, and would require a rebuild of the deployment (not practical!). Thankfully, useRouter has the solution for us. It stores the url in a query, so we can access it from the file [settlement].js endpoint as router.query.settlement